+++ CHANGELOG.md
@@ -2,11 +2,16 @@
 
 ## 1.0.0 (In Development)
 
+### 1.0.0-alpha.4 (In Development)
+
+- Provide `onSuccessResponse` & `onErrorResponse` request callbacks
+- Allow to pass `deps` dependencies to inner `useEffect` call
+
 ### 1.0.0-alpha.3 (2020-03-24)
 
 - Provide [storybook](https://react-sadness.now.sh) to show how to use
   `react-sadness` library
-- Fix `onReload` & `onUpdate` callbacks
+- Fix `onReload` & `onUpdate` response callbacks
 
 ### 1.0.0-alpha.2 (2020-03-23)